requirements.txt
# Runtime
Flask==3.0.3
psycopg2-binary==2.9.9
Pillow==10.3.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.3

# Security (Fase 2)
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
flask-talisman==1.1.0

# Testing
pytest==8.2.0
pytest-flask==1.3.0
